Stop Session Dialogue flow from app block flow and stop focus and reset the timer
Enable both the application block and website url script. App-launch block is not working when script is running. find the solution to run both the functionality parallel.
Design the One mintue block screen and set it when the break start.
Setup the data on the Current session screen and perform the stop session functionality here too.
Checking the Database migration issue. Fixed with the versioning. (Still need to check)
Prepare the installer file for build update